Upload all gzippedFiles that are not found in the distFiles
With the use of the `keep` option in the gzip plugin you will end
up with a set of gzippedFiles that are not in the context's
distFiles. These files will not be uploaded to S3.

I've modified this plugin so that the files to be uploaded is now
the concatenation of distFiles and gzippedFiles. For the non-keep
users the effective result is a no-op, but for people who use keep
this means their gzipped files will actually make it up to S3.